+++ phpunit/PHPUnit_Framework_TestCase.php
@@ -1,0 +1,16 @@
+<?php
+
+/*
+ * This file is part of the Symfony CMF package.
+ *
+ * (c) 2011-2017 Symfony CMF
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+use PHPUnit\Framework\TestCase;
+
+if (class_exists(TestCase::class)) {
+    class_alias(TestCase::class, 'PHPUnit_Framework_TestCase');
+}