+++ src/Phpunit/DatabaseTestListener.php
@@ -14,8 +14,9 @@ namespace Symfony\Cmf\Component\Testing\Phpunit;
 use Symfony\Component\Process\ProcessBuilder;
 use Symfony\Component\Process\PhpExecutableFinder;
 use Doctrine\Common\DataFixtures\Purger;
+use PHPUnit\Framework\BaseTestListener;
 
-class DatabaseTestListener implements \PHPUnit_Framework_TestListener
+class DatabaseTestListener extends BaseTestListener
 {
     protected static $currentSuite;
     private $processBuilder;
@@ -38,30 +39,6 @@ class DatabaseTestListener implements \PHPUnit_Framework_TestListener
         }
     }
 
-    public function addError(\PHPUnit_Framework_Test $test, \Exception $e, $time)
-    {
-    }
-
-    public function addFailure(\PHPUnit_Framework_Test $test, \PHPUnit_Framework_AssertionFailedError $e, $time)
-    {
-    }
-
-    public function addWarning(\PHPUnit_Framework_Test $test, \PHPUnit_Framework_Warning $e, $time)
-    {
-    }
-
-    public function addIncompleteTest(\PHPUnit_Framework_Test $test, \Exception $e, $time)
-    {
-    }
-
-    public function addSkippedTest(\PHPUnit_Framework_Test $test, \Exception $e, $time)
-    {
-    }
-
-    public function addRiskyTest(\PHPUnit_Framework_Test $test, \Exception $e, $time)
-    {
-    }
-
     public function startTest(\PHPUnit_Framework_Test $test)
     {
         switch (static::$currentSuite->getName()) {
@@ -81,10 +58,6 @@ class DatabaseTestListener implements \PHPUnit_Framework_TestListener
         }
 
         $purger->purge();
-    }
-
-    public function endTest(\PHPUnit_Framework_Test $test, $time)
-    {
     }
 
     public function startTestSuite(\PHPUnit_Framework_TestSuite $suite)